+++ src/main/kotlin/com/github/sam0delkin/intellijpsa/language/php/PhpPsaExtension.kt
@@ -1,10 +1,12 @@
 package com.github.sam0delkin.intellijpsa.language.php
 
 import com.github.sam0delkin.intellijpsa.extension.extensionPoints.PsaExtension
+import com.github.sam0delkin.intellijpsa.language.php.eventSubscriber.PsaXdebugManagerListener
 import com.github.sam0delkin.intellijpsa.language.php.model.PhpInfoModel
 import com.github.sam0delkin.intellijpsa.language.php.services.PhpPsaManager
 import com.github.sam0delkin.intellijpsa.language.php.settings.PhpPsaSettings
 import com.github.sam0delkin.intellijpsa.settings.Settings
+import com.intellij.openapi.Disposable
 import com.intellij.openapi.actionSystem.ActionGroup
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.service
@@ -13,18 +15,21 @@ import com.intellij.ui.components.JBCheckBox
 import com.intellij.ui.dsl.builder.Cell
 import com.intellij.ui.dsl.builder.Panel
 import com.intellij.ui.dsl.builder.selected
-import kotlinx.serialization.decodeFromString
+import com.intellij.util.messages.MessageBusConnection
+import com.intellij.xdebugger.XDebuggerManager
 import kotlinx.serialization.json.Json
 
-class PhpPsaExtension : PsaExtension {
+class PhpPsaExtension :
+    PsaExtension,
+    Disposable {
     private lateinit var enabled: Cell<JBCheckBox>
     private lateinit var debugTypeProvider: Cell<JBCheckBox>
+    private var connection: MessageBusConnection? = null
 
     override fun configure(
         panel: Panel,
         project: Project,
     ) {
-        val settings = project.service<PhpPsaSettings>()
         panel.group("PHP") {
             row("Enabled") {
                 enabled = checkBox("")
@@ -62,24 +67,44 @@ class PhpPsaExtension : PsaExtension {
         project: Project,
         info: String,
     ) {
+        val psaManager = project.service<PhpPsaManager>()
+        val settings = project.service<Settings>()
+        val phpSettings = psaManager.getSettings()
+
         try {
-            val phpInfo = Json.Default.decodeFromString<PhpInfoModel>(info)
-            val psaManager = project.service<PhpPsaManager>()
-            val settings = project.service<Settings>()
-            val phpSettings = psaManager.getSettings()
+            val json =
+                Json {
+                    ignoreUnknownKeys = true
+                }
+            val phpInfo = json.decodeFromString<PhpInfoModel>(info)
 
             if (!phpSettings.enabled) {
                 return
             }
 
             phpSettings.supportsTypeProviders = phpInfo.supportsTypeProviders ?: false
+            phpSettings.toStringValueFormatter = phpInfo.toStringValueFormatter
 
             if (phpSettings.supportsTypeProviders) {
                 ApplicationManager.getApplication().invokeLater {
                     psaManager.updateTypeProviders(settings, phpSettings, project)
                 }
             }
-        } catch (e: Throwable) {
+
+            if (
+                !phpSettings.enabled ||
+                null === phpSettings.toStringValueFormatter
+            ) {
+                return
+            }
+
+            connection?.disconnect()
+            connection = project.messageBus.connect()
+            connection?.subscribe(
+                XDebuggerManager.TOPIC,
+                PsaXdebugManagerListener(project),
+            )
+        } catch (_: Throwable) {
             return
         }
     }
@@ -88,5 +113,9 @@ class PhpPsaExtension : PsaExtension {
         project: Project,
         actionGroup: ActionGroup,
     ) {
+    }
+
+    override fun dispose() {
+        this.connection?.disconnect()
     }
 }